Move `CoffeeNetProfileAutoConfiguration` a level up
In prepartion for WebFlux support, we move the configuration for the
`CoffeeNetProfileMapper` into the parent package.
