+++ coffeenet-autoconfigure/src/main/java/rocks/coffeenet/autoconfigure/security/CoffeeNetProfileAutoConfiguration.java
@@ -1,8 +1,9 @@
-package rocks.coffeenet.autoconfigure.security.servlet;
+package rocks.coffeenet.autoconfigure.security;
 
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnWebApplication;
 import org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration;
@@ -12,10 +13,13 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.springframework.context.annotation.Primary;
 
+import org.springframework.security.authentication.DefaultAuthenticationEventPublisher;
 import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
 
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+
+import rocks.coffeenet.autoconfigure.security.servlet.CoffeeNetProfileArgumentResolver;
 
 import rocks.coffeenet.platform.domain.profile.CoffeeNetProfile;
 import rocks.coffeenet.platform.domain.profile.DelegatingPrincipalCoffeeNetProfileMapper;
@@ -31,9 +35,8 @@ import java.util.List;
  * @since  2.0.0
  */
 @Configuration(proxyBeanMethods = false)
-@AutoConfigureAfter(SecurityAutoConfiguration.class)
-@ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.SERVLET)
-@ConditionalOnBean(name = AbstractSecurityWebApplicationInitializer.DEFAULT_FILTER_NAME)
+@AutoConfigureAfter({ SecurityAutoConfiguration.class })
+@ConditionalOnClass(DefaultAuthenticationEventPublisher.class)
 @Import(FallbackProfileMapperConfiguration.class)
 public class CoffeeNetProfileAutoConfiguration {
 
@@ -48,7 +51,11 @@ public class CoffeeNetProfileAutoConfiguration {
     }
 
     @Configuration(proxyBeanMethods = false)
-    @ConditionalOnBean(PrincipalCoffeeNetProfileMapper.class)
+    @ConditionalOnWebApplication(type = ConditionalOnWebApplication.Type.SERVLET)
+    @ConditionalOnBean(
+        value = { PrincipalCoffeeNetProfileMapper.class },
+        name = AbstractSecurityWebApplicationInitializer.DEFAULT_FILTER_NAME
+    )
     static class CoffeeNetProfileArgumentResolverConfigurer implements WebMvcConfigurer {
 
         private final PrincipalCoffeeNetProfileMapper mapper;
