+++ coffeenet-autoconfigure/src/main/java/rocks/coffeenet/autoconfigure/security/FallbackProfileMapperConfiguration.java
@@ -1,11 +1,9 @@
-package rocks.coffeenet.autoconfigure.security.servlet;
+package rocks.coffeenet.autoconfigure.security;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-
-import rocks.coffeenet.autoconfigure.security.DefaultPrincipalCoffeeNetProfileMapper;
 
 import rocks.coffeenet.platform.domain.profile.PrincipalCoffeeNetProfileMapper;
