+++ coffeenet-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,8 +1,8 @@
 # Auto-configuration
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-rocks.coffeenet.autoconfigure.security.oauth2.servlet.OAuth2SecurityAutoConfiguration,\
-rocks.coffeenet.autoconfigure.security.servlet.CoffeeNetProfileAutoConfiguration,\
-rocks.coffeenet.autoconfigure.security.servlet.CoffeeNetSecurityAutoConfiguration
+rocks.coffeenet.autoconfigure.security.CoffeeNetProfileAutoConfiguration,\
+rocks.coffeenet.autoconfigure.security.CoffeeNetSecurityAutoConfiguration,\
+rocks.coffeenet.autoconfigure.security.oauth2.servlet.OAuth2SecurityAutoConfiguration
 
 # Injected security configurers
 org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer=\
